# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/pipeline/dispatch": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Appends a user message to the given session and streams the pipeline response as NDJSON. Each line is a WireEvent JSON object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/x-ndjson"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Dispatch message",
                "parameters": [
                    {
                        "description": "Dispatch request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.dispatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "NDJSON stream; one event per line",
                        "schema": {
                            "$ref": "#/definitions/pipeline.WireEvent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/pipeline/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the assembled system prompt and chat-message slice that would be sent to the provider, without persisting the new user message or calling the LLM. Use to debug prompt composition.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Preview pipeline input",
                "parameters": [
                    {
                        "description": "Preview request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pipeline.previewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pipeline.previewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/plugins/{name}": {
            "get": {
                "description": "Returns info for a single plugin driver by name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugins"
                ],
                "summary": "Get plugin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plugin name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/plugins.PluginDriverInfo"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/plugins/{name}/capabilities": {
            "get": {
                "description": "Returns the capability set advertised by a plugin driver",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "plugins"
                ],
                "summary": "Get plugin capabilities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plugin name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pipeline.WireEvent": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pipeline.dispatchRequest": {
            "type": "object",
            "required": [
                "content",
                "session_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "no_store": {
                    "type": "boolean"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "pipeline.previewRequest": {
            "type": "object",
            "required": [
                "session_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "pipeline.previewResponse": {
            "type": "object",
            "properties": {
                "est_accuracy": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pipeline.previewResponseMessage"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "system": {
                    "type": "string"
                },
                "system_usage": {
                    "$ref": "#/definitions/pipeline.previewUsage"
                },
                "tool_count": {
                    "type": "integer"
                },
                "total": {
                    "$ref": "#/definitions/pipeline.previewUsage"
                }
            }
        },
        "pipeline.previewResponseMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/pipeline.previewUsage"
                }
            }
        },
        "pipeline.previewUsage": {
            "type": "object",
            "properties": {
                "bytes": {
                    "type": "integer"
                },
                "est_tokens": {
                    "type": "integer"
                },
                "runes": {
                    "type": "integer"
                }
            }
        },
        "plugins.PluginDriverInfo": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "enabled": {
                    "type": "boolean"
                },
                "env": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "max_port": {
                    "type": "integer"
                },
                "min_port": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "timeout": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}